I'm giving up on a principled build for now
I will build parser and scanner manually when I need to and check in the changes. I am so not in the mood to figure out to convince GNU Make and the VCV Rack build config for it.

Hacked out much (not all) of the guts of Basically.cpp, and now it actually computes something.