Adding Randomize revealed many problems
Undo/redo had a bunch of crashing problems, which ended up being about old indexes being used when the size of "circles" has changed. Instead of deleting and adding Circle objects, the vector is always 16 in size, and we just copy to it.

Realized that the way the Initialize and Randomize menu items work, I don't need to make undo/redo objects for them.

Now must explicitly save/read the "present" boolean in the text, otherwise it fails to parse.